backend/Transcript/requirements.txt
youtube-transcript-api
fastapi
pydantic
google-generativeai
uvicorn
